Implement ducks pattern, update tests
