--- src/actions/counter.js
@@ -1,13 +1,0 @@
-import { INCREMENT_COUNTER, DECREMENT_COUNTER } from '../constants';
-
-export function increment() {
-  return {
-    type: INCREMENT_COUNTER,
-  };
-}
-
-export function decrement() {
-  return {
-    type: DECREMENT_COUNTER,
-  };
-}
