+++ src/ducks/counter.js
@@ -1,9 +1,30 @@
-import { INCREMENT_COUNTER, DECREMENT_COUNTER, LOGOUT_USER } from '../constants';
 import { fromJS } from 'immutable';
+import { LOGOUT_USER } from './session.js';
+
+// constants
+
+export const INCREMENT_COUNTER = 'App/counter/INCREMENT_COUNTER';
+export const DECREMENT_COUNTER = 'App/counter/DECREMENT_COUNTER';
 
 const INITIAL_STATE = fromJS({
   count: 0,
 });
+
+// action creators
+
+export function increment() {
+  return {
+    type: INCREMENT_COUNTER,
+  };
+}
+
+export function decrement() {
+  return {
+    type: DECREMENT_COUNTER,
+  };
+}
+
+// reducer
 
 function counterReducer(state = INITIAL_STATE, action = {}) {
   switch (action.type) {
