+++ src/store/configure-store.js
@@ -4,7 +4,7 @@ import persistState from 'redux-localstorage';
 import thunk from 'redux-thunk';
 import promiseMiddleware from '../middleware/promise-middleware';
 import logger from './logger';
-import rootReducer from '../reducers';
+import rootReducer from '../ducks';
 
 function configureStore(initialState) {
   const store = compose(
@@ -43,8 +43,8 @@ function _getEnhancers() {
 
 function _enableHotLoader(store) {
   if (__DEV__ && module.hot) {
-    module.hot.accept('../reducers', () => {
-      const nextRootReducer = require('../reducers');
+    module.hot.accept('../ducks', () => {
+      const nextRootReducer = require('../ducks');
       store.replaceReducer(nextRootReducer);
     });
   }
